Modify the stacked chart component to use dynamic labels
