+++ src/StackChartContainer.js
@@ -1,13 +1,12 @@
-import AirflowStackedChart from './AirflowStackedChart'
-import TemperatureStackedChart from "./TemperatureStackedChart";
+import StackedChart from "./StackedChart";
 
 export default function StackChartContainer() {
     return (
         <div className = "border">
 	    <p className = "title">Zone Temperature</p>
-            <TemperatureStackedChart />
+            <StackedChart type={'temperature'} />
 	    <p style={{borderTop:"1px solid black"}} className = "title"><br/>Zone Airflow</p>
-            <AirflowStackedChart />
+            <StackedChart type={'airflow'} />
         </div>
     );
 }
